feature: Calculate amount of transactions on a wallet

+++ src/main/java/com/altamirobruno/save_my_money/dto/TransactionDTO.java
@@ -2,13 +2,14 @@ package com.altamirobruno.save_my_money.dto;
 
 import com.altamirobruno.save_my_money.enums.TransactionType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 public record TransactionDTO(
         UUID id,
         String name,
-        Float value,
+        BigDecimal value,
         String description,
         LocalDate date,
         String installment,
